+++ site/templates/display_page.html
@@ -23,7 +23,7 @@
                      <div class="card hoverable">
                         <div class="card-image waves-effect waves-block waves-light valign-wrapper" style="height:450px">
                            {% if rnr_entity['rights'] != None %}
-                              <img class="activator" src = {{rnr_entity["path_to_image"] }}>
+                              <img class="activator" src={{ rnr_entity['path_to_image'] }}>
                            {%endif%}
                       </div>
                         <div class="card-content">
